+++ lib/widgets/drawer_nav.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class DrawerNavigationItem extends StatelessWidget {
+  final IconData iconData;
+  final String title;
+  final bool selected;
+  final Function() onTap;
+  const DrawerNavigationItem({
+    Key? key,
+    required this.iconData,
+    required this.title,
+    required this.selected,
+    required this.onTap,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(30.0),
+      ),
+      leading: Icon(iconData),
+      onTap: onTap,
+      title: Text(title),
+      selectedTileColor: Colors.blueAccent.shade100,
+      selected: selected,
+      selectedColor: Colors.black87,
+    );
+  }
+}
